feat: extract @module-federation/data-fetch from bridge-react
