+++ packages/bridge/bridge-react/src/data-fetch.ts
@@ -1,3 +1,13 @@
+export type {
+  DataFetchParams,
+  NoSSRRemoteInfo,
+  CollectSSRAssetsOptions,
+  CreateLazyComponentOptions,
+  CacheStatus,
+  CacheStatsInfo,
+  PrefetchOptions,
+} from './lazy';
+
 export {
   ERROR_TYPE,
   createLazyComponent,
@@ -13,17 +23,7 @@ export {
   revalidateTag,
   clearStore,
   prefetch,
+  flushDataFetch,
 } from './lazy';
 
 export { lazyLoadComponentPlugin } from './plugins/lazy-load-component-plugin';
-export { flushDataFetch } from './lazy/utils';
-
-export type {
-  DataFetchParams,
-  NoSSRRemoteInfo,
-  CollectSSRAssetsOptions,
-  CreateLazyComponentOptions,
-  CacheStatus,
-  CacheStatsInfo,
-  PrefetchOptions,
-} from './lazy';
