+++ packages/bridge/bridge-react/src/lazy/data-fetch/index.ts
@@ -1,16 +1,1 @@
-export { callDataFetch } from './call-data-fetch';
-export { injectDataFetch } from './inject-data-fetch';
-export {
-  CacheSize,
-  CacheTime,
-  configureCache,
-  generateKey,
-  cache,
-  revalidateTag,
-  clearStore,
-} from './cache';
-
-export type { CacheStatus, CacheStatsInfo } from './cache';
-
-export type { PrefetchOptions } from './prefetch';
-export { prefetch } from './prefetch';
+export * from '@module-federation/data-fetch/data-fetch-utils';
