+++ packages/bridge/vue3-bridge/src/index.ts
@@ -2,3 +2,22 @@ export { createBridgeComponent } from './provider';
 export type { ProviderFnParams } from './provider';
 export { createRemoteComponent, createRemoteAppComponent } from './create';
 export type { RenderFnParams } from '@module-federation/bridge-shared';
+
+export {
+  autoFetchDataPlugin,
+  cache,
+  clearStore,
+  configureCache,
+  generateKey,
+  prefetch,
+  revalidateTag,
+  CacheSize,
+  CacheTime,
+} from '@module-federation/data-fetch';
+
+export type {
+  CacheStatus,
+  CacheStatsInfo,
+  DataFetchParams,
+  PrefetchOptions,
+} from '@module-federation/data-fetch';
