+++ packages/data-fetch/src/env.d.ts
@@ -1,0 +1,8 @@
+import type { ModuleFederation } from '@module-federation/runtime';
+
+declare global {
+  var __FEDERATION__: {
+    moduleInfo?: unknown;
+    __INSTANCES__: ModuleFederation[];
+  };
+}
