+++ packages/data-fetch/src/logger.ts
@@ -1,0 +1,6 @@
+import { createLogger } from '@module-federation/sdk';
+import { PLUGIN_IDENTIFIER } from './constant';
+
+const logger = createLogger(PLUGIN_IDENTIFIER);
+
+export default logger;
