+++ packages/data-fetch/src/size-limited-cache.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@rstest/core';
-import { SizeLimitedCache } from './size-limited-cache';
+import { SizeLimitedCache } from '../src/size-limited-cache';
 
 describe('SizeLimitedCache', () => {
   it('evicts the least recently used entries by total size', () => {
